♻️ refactor(ModelSelect): move NewModelBadge to shared components
- Move NewModelBadge from src/features/ModelSelect/components to src/components/ModelSelect
- Reuse NewModelBadge in ModelItemRender component
- Remove internal margin styling, let consumers control spacing
- Update import paths in ModelItem and ImageModelItem
