+++ src/components/ModelSelect/index.tsx
@@ -21,6 +21,8 @@ import { Flexbox } from 'react-layout-kit';
 import { AiProviderSourceType } from '@/types/aiProvider';
 import { formatTokenNumber } from '@/utils/format';
 
+import NewModelBadge from './NewModelBadge';
+
 export const TAG_CLASSNAME = 'lobe-model-info-tags';
 
 const useStyles = createStyles(({ css, token }) => ({
@@ -179,6 +181,7 @@ interface ModelItemRenderProps extends ChatModelCard {
 
 export const ModelItemRender = memo<ModelItemRenderProps>(({ showInfoTag = true, ...model }) => {
   const { mobile } = useResponsive();
+
   return (
     <Flexbox
       align={'center'}
@@ -202,6 +205,7 @@ export const ModelItemRender = memo<ModelItemRenderProps>(({ showInfoTag = true,
         <Text style={mobile ? { maxWidth: '60vw', overflowX: 'auto', whiteSpace: 'nowrap' } : {}}>
           {model.displayName || model.id}
         </Text>
+        <NewModelBadge releasedAt={model.releasedAt} />
       </Flexbox>
       {showInfoTag && <ModelInfoTags {...model} />}
     </Flexbox>
